+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.1.0 (#unreleased)
 
 - Feature [#468](https://github.com/SimformSolutionsPvtLtd/audio_waveforms/pull/468) - Add macOS support
+- Fixed [#491](https://github.com/SimformSolutionsPvtLtd/audio_waveforms/issues/491) - Update totalSamples calculation to account for bytes per sample
 
 ## 2.0.2